services/products/requirements.txt
Django==4.2.2
djangorestframework
djongo
gunicorn
requests
